moved test for import_module()
